+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - Requires php `7.1`.
+- Removed deprecated symbols.
 
 ## [1.1.0] - 2017-12-20